requirements.txt
gradio
requests
python-dotenv
openai
tavily-python
